fix: use joinValues()
