+++ packages/client/src/v2-events/features/events/Search/utils.test.ts
@@ -11,10 +11,10 @@
 
 import {
   EventStatus,
+  joinValues,
   tennisClubMembershipEvent
 } from '@opencrvs/commons/client'
 import { FIELD_SEPARATOR } from '@client/v2-events/components/forms/utils'
-import { joinValues } from '@client/v2-events/utils'
 import {
   getAdvancedSearchFieldErrors,
   flattenFieldErrors,
