+++ packages/commons/src/utils.ts
@@ -16,3 +16,17 @@ export function getOrThrow<T>(x: T, message: string) {
 
   return x
 }
+
+/**
+ *
+ * Joins defined values using a separator and trims the result
+ */
+export function joinValues(
+  values: Array<string | undefined | null>,
+  separator = ' '
+) {
+  return values
+    .filter((value) => !!value)
+    .join(separator)
+    .trim()
+}
